fix(billing): use BigInteger for notification tables and fix bulk_insert in migration

+++ mona-auth/app/models.py
@@ -1,7 +1,7 @@
 import enum
 from datetime import datetime
 
-from sqlalchemy import DateTime, Enum, ForeignKey, Integer, String, UniqueConstraint, func
+from sqlalchemy import BigInteger, DateTime, Enum, ForeignKey, Integer, String, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -152,7 +152,7 @@ class AppConfig(Base):
 class Notification(Base):
     __tablename__ = "notifications"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     title: Mapped[str] = mapped_column(String(128), nullable=False)
     body: Mapped[str] = mapped_column(String(512), nullable=False)
     type: Mapped[str] = mapped_column(String(32), nullable=False)
@@ -168,11 +168,11 @@ class NotificationRead(Base):
     __tablename__ = "notification_reads"
     __table_args__ = (UniqueConstraint("user_id", "notification_id"),)
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     user_id: Mapped[int] = mapped_column(
-        Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True
+        BigInteger, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True
     )
     notification_id: Mapped[int] = mapped_column(
-        Integer, ForeignKey("notifications.id", ondelete="CASCADE"), nullable=False, index=True
+        BigInteger, ForeignKey("notifications.id", ondelete="CASCADE"), nullable=False, index=True
     )
     read_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now())
